% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bain_methods.R
\name{bain}
\alias{bain}
\title{Bayes factors for informative hypotheses}
\usage{
bain(x, hypothesis, ...)
}
\arguments{
\item{x}{An object used to select a method for computing Bayes factors.
Currently, methods are available for:
\itemize{
\item Named vectors of estimates
\item lm objects
}}

\item{hypothesis}{A character string, containing a Bain hypothesis (see Details).}

\item{...}{Arguments passed to and from other functions.}
}
\value{
Returns a list that contains hypothesis testing results (i.e., Bayes
factors (BFs), relative fit (f), relative complexity (c), posterior model
probabilities (PMPs)), the approximated posterior and prior covariance
matrices, the approximated posteriors means, and the fraction (b).
}
\description{
This function computes approximated adjusted fractional Bayes factors
(AAFBFs) for equality, inequality, and about equality constrained
hypotheses. The function can also be used for computing the AAFBF in the
case of multiple groups.
}
\details{
The vector "estimate" and the matrix (or list) "Sigma" should be obtained
before evaluating hypotheses in Bain. For examples, see Gu, Mulder, and
Hoijtink (2017) and Hoijtink, Gu, and Mulder (unpublished) and under the
MGBain button at http://informative-hypotheses.sites.uu.nl/ were also the
papers and other information can be found.

The length of "estimate" should be equal to grouppara * P + jointpara, where
P denotes the number of groups or the length of list "Sigma".

"ERr" is used to specify equality constraints, and "IRr" is used to specify
inequality constraints. The general form of the constraints is ER * t^T = r
and IR * t^T = r where t is a vector containing the estimates, ER is a
length of t by number of equality constraints matrix, IR is a length of t by
number of inequality constraints matrix, and r (used with ER) has a length
equal to the number of equality constraints, and r (used with IR) has a
length equal to the number of inequality constraints. ER, IR, and r contain
real numbers.

%%For example, an equality constraint t1=t2+1 can be specified as
ER*(t1,t2)^T = r, where ER = (1,-1), and r = 1. %%Then ERr = (ER,r) =
(1,-1,1), which is provided to Bain using the command
ERr<-matrix(c(1,-1,1),nrow=1,ncol=3%%,byrow = TRUE).

%%Another example, if the inequality constraints are t1 > t2 and t2 > t3
then IR * (t1,t2,t3)^T = r where IR= \cr %%1 -1 0 \cr %%0 \sspace 1 -1 \cr
%%and r = (0 0). %%Then IRr =\cr %%1 -1 0 0 \cr %%0 1 -1 0 \cr %%which is
provided to Bain using the command:\cr
%%IRr<-matrix(c(1,-1,0,0,0,1,-1,0),nrow=2,ncol=4,byrow = TRUE)

Informative hypotheses should adhere to the following simple syntax:
\itemize{
\item Competing hypotheses are separated by ";".  Thus, "a=b;a>b" means that
H1: a=b, and H2: a>b.
\item Each individual hypothesis consists of a (series of) (in)equality
constraint(s). Every single (in)equality constraint is of the form "R1*mu1 +
R2*mu2+... = r", where capital Rs refer to numeric scaling constants, must
refer to the names of parameters in the model, and the lower case r refers
to a constant. Standard mathematical simplification rules apply; thus,
"R1*mu1 = R2*mu2" is equivalent to "R1*mu1 - R2*mu2 = 0".
\item Multiple unrelated constraints within one hypothesis can be chained by
"&". Thus, "a=b&c=d" means that H1: a=b AND c=d.
\item Multiple related constraints within one hypothesis can be chained by
repeating the (in)equality operators "=", "<", or ">". Thus, "a<b<c" means
that H1: a < b AND b < c.
\item Parameters can be grouped by placing them in a parenthesized, comma
separated list. Thus, "(a,b)>c" means that H1: a > c AND b > c.  Similarly,
"(a,b)>(c,d)" means that H1: a > c AND b > c AND b > c AND b > d.
}
}
\examples{
\dontrun{
#One group:
#Example 1:
#Hypothesis
#H1: theta1>theta2>theta3   #coefficients in regression model

#Input
estimate<-c(3,2,1)  #estimates of coefficients
Sigma<-matrix(c(3,0,0,0,2,0,0,0,1),3,3,byrow = TRUE) #covariance matrix of coefficients

n<-50 #samples size

#H1
ERr<-NULL
IRr<-matrix(c(1,-1,0,0,0,1,-1,0),nrow=2,ncol=4,byrow = TRUE)

res<-Bain(estimate=estimate,grouppara=0,jointpara=3,Sigma=Sigma,n=n,ERr,IRr) #run
#Results are printed.
#Results for fit, complexity, Bayes factor, and PMPs are saved in "res":

plot(res)
#Results for PMPs are plotted.


#Multiple groups
#Example 2
#t test:
#Hypotheses:
#H1: theta1=theta2   #group means
#H2: theta1>theta2
#H3: theta1<theta2

#Input
estimate<-c(0,0)         #Estimates of group means theta1 and theta2.

cov1<-matrix(c(.5),1,1)
cov2<-matrix(c(.1),1,1)
Sigma<-list(cov1,cov2) #List of variances of group means

n<-c(22,37)           #samplesize

#H1:
ERr1<-matrix(c(1,-1,0),nrow=1,ncol=3,byrow = TRUE)
IRr1<-NULL

#H2
ERr2<-NULL
IRr2<-matrix(c(1,-1,0),nrow=1,ncol=3,byrow = TRUE)

#H3
ERr3<-NULL
IRr3<-matrix(c(-1,1,0),nrow=1,ncol=3,byrow = TRUE)

res<-Bain(estimate,Sigma,grouppara=1,jointpara=0,n=n,ERr1,IRr1,ERr2,IRr2,ERr3,IRr3) #run
#Results are printed.
#Results for fit, complexity, Bayes factor, and PMPs are also saved in "res":

plot(res)
#Results for PMPs are plotted.
}
}
\references{
Gu, X., Mulder, J., and Hoijtink, H. (2017). Approximated
adjusted fractional Bayes factors: A general method for testing informative
hypotheses. British Journal of Mathematical and Statistical Psychology.

Hoijtink, H., Gu, X., and Mulder, J. (unpublished). Multiple group Bayesian
evaluation of informative hypotheses.
}
\author{
Caspar van Lissa, Xin Gu, Herbert Hoijtink, Joris Mulder
}
\keyword{htest}
