% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bain_sensitivity.R
\name{bain_sensitivity}
\alias{bain_sensitivity}
\title{Sensitivity analysis for bain}
\usage{
bain_sensitivity(x, hypothesis, fractions = 1, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{hypothesis}{PARAM_DESCRIPTION}

\item{fractions}{PARAM_DESCRIPTION, Default: 1}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Conducts a sensitivity analysis for \code{\link[bain]{bain}},
see details.
\code{fractions} argument, and returns a list of bain results objects.
}
\details{
The Bayes factor for equality constraints is sensitive to a
scaling factor applied to the prior distribution. The \code{fraction}
argument adjusts this scaling factor. The function \code{bain_sensitivity}
is a wrapper for \code{\link[bain]{bain}}, which accepts a vector for the
\code{fractions} argument, and returns a list of bain results objects.
A table with a sensitivity analysis for specific statistics can be obtained
using the \code{summary()} function, which accepts the argument
\code{summary(which_stat = ...)}. The available statistics are elements of
the \code{$fit} table (Fit_eq, Com_eq, Fit_in, Com_in, Fit, Com, BF, PMPa,
and PMPb), and elements of the
\code{BFmatrix}, which can be accessed by matrix notation, e.g.:
\code{summary(bain_sens, which_stat = "BFmatrix[1,2]")}.
}
\examples{
res <- t_test(extra ~ group, data = sleep)
bain(res, hypothesis = "group1 - group2 = 1.5")
bain_sens <- bain_sensitivity(res, hypothesis = "group1-group2 = 1.5",
                              fractions = c(1,2,3))
summary(bain_sens, which_stat = "BF")
summary(bain_sens, which_stat = "BFmatrix[1,1]")
}
