% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrices.R
\name{constraint_to_row}
\alias{constraint_to_row}
\title{Constraint to row}
\usage{
constraint_to_row(varnames, hyp)
}
\arguments{
\item{hyp}{Character. A BAIN (in)equality constraint formatted as equation,
with all terms on the left hand side.}
}
\value{
Numeric vector.
}
\description{
Evaluate a BAIN constraint, which has been formatted as an equation, with
all terms moved to the left hand side, and return a single row of a BAIN
(in)equality constraint matrix.
}
\examples{

constraint_to_row(varnames = c("a", "b", "c", "d"), hyp = "+1*a-1*b+0=")
constraint_to_row(c("a", "b", "c", "d"), "+1*b-1*c+0=")
constraint_to_row(c("a", "b", "c", "d"), "-+0+1*c-1*d=")

}
\keyword{internal}
