% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\title{Get descriptive statistics from a bain model}
\usage{
descriptives(x, ci = 0.95)
}
\arguments{
\item{x}{An object of class 'bain'.}

\item{ci}{Numeric. Percentage for two-tailed confidence interval, defaults to
.95.}
}
\value{
A data.frame.
}
\description{
Get descriptive statistics from a bain model
}
