% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrices.R
\name{expand_compound_constraints}
\alias{expand_compound_constraints}
\title{Expand compound constraints}
\usage{
expand_compound_constraints(hyp)
}
\arguments{
\item{hyp}{Character. A BAIN (in)equality constraint}
}
\value{
A character vector with one element for each simple constraint
}
\description{
Takes a compound BAIN constraint, with multiple (in)equality operators, and
expands it into simple constraints.
}
\keyword{internal}
