% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrices.R
\name{expand_parentheses}
\alias{expand_parentheses}
\title{Expand parentheses}
\usage{
expand_parentheses(hyp)
}
\arguments{
\item{hyp}{Character. An (in)equality constraint}
}
\value{
A character vector with one element for each simple constraint
}
\description{
Takes a constraint with parentheses containing a "vector" of parameter
labels, and recursively expands the parenthesized "vectors".
}
\keyword{internal}
