% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrices.R
\name{flip_inequality}
\alias{flip_inequality}
\title{Flip inequality}
\usage{
flip_inequality(hyp)
}
\arguments{
\item{hyp}{Character. An (in)equality constraint}
}
\value{
Character
}
\description{
Takes a constraint and flips elements on both sides of any "<"
operators so that only the ">" operator is used to define inequality
constraints.
}
\keyword{internal}
