% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates.R
\name{get_estimates}
\alias{get_estimates}
\title{Get estimates from a model object}
\usage{
get_estimates(x, ...)
}
\arguments{
\item{x}{A model object for which a \code{\link{bain}} method exists.}

\item{...}{Parameters passed to and from other functions.}
}
\value{
A named numeric vector.
}
\description{
Get estimates from a model object, the way the
\code{\link{bain}} function will. This convenience function allows you to see
that coefficients are properly extracted, note how their names will be parsed
by bain, and inspect their values.
}
\keyword{internal}
