% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates.R
\name{label_estimates}
\alias{label_estimates}
\title{Label estimates from a model object}
\usage{
label_estimates(x, labels, ...)
}
\arguments{
\item{x}{A model object for which a \code{\link{bain}} method exists.}

\item{labels}{Character vector. New labels (in order of appearance) for the
model object in \code{x}. If you are unsure what the estimates in \code{x}
are, first run \code{\link{get_estimates}}.}

\item{...}{Parameters passed to and from other functions.}
}
\value{
A model object of the same class as x.
}
\description{
Label estimates from a model object, before passing it on to the
\code{\link{bain}} function.
}
\examples{
\dontrun{
# Example 1
m_tt <- t.test(iris$Sepal.Length[1:20], iris$Sepal.Length[21:40])
m_tt <- label_estimates(m_tt, c("a", "b"))
# Example 2
m_lm <- lm(Sepal.Length ~., iris)
m_lm <- label_estimates(m_lm, labels = c("a", "b", "c", "d", "e", "f"))
}
}
\seealso{
get_estimates bain
}
