% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrices.R
\name{order_terms}
\alias{order_terms}
\title{Order terms}
\usage{
order_terms(hyp)
}
\arguments{
\item{hyp}{Character. A BAIN (in)equality constraint formatted as equation.}
}
\value{
Character
}
\description{
Moves all terms on the right hand side of a BAIN constraint, which has been
formatted as an equation, to the left hand side.
}
\examples{

order_terms("+2+1*a-1=+1*b-2")
order_terms(hyp = "+1*a=-2")
order_terms(hyp = "+1*a=-2+3")
order_terms(hyp = "+1*a=+1*b")
order_terms(hyp = "-2+1*a=+1*b")

}
\keyword{internal}
