% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameCFA.R
\docType{data}
\name{sesameCFA}
\alias{sesameCFA}
\title{Simulated Sesame Street Data for CFA}
\format{A data frame with 240 rows and 15 variables.}
\usage{
data(sesameCFA)
}
\description{
A simulated data set inspired by the Sesame Street data set from:
Stevens, J. P. (1996). Applied Multivariate Statistics for the Social
Sciences. Mahwah NJ: Lawrence Erlbaum.
}
\details{
The number knowledge of children is measured before watching Sesame Street,
after one year, and a follow up measurement after two years.

\describe{
  \item{age}{Age of the child in months}
  \item{pea}{Peabody mental age score of the child; the higher the
  score the higher the mental age}
  \item{Bb}{Factor B item 1}
  \item{Bl}{Factor B item 2}
  \item{Bf}{Factor B item 3}
  \item{Bn}{Factor B item 4}
  \item{Br}{Factor B item 5}
  \item{Bc}{Factor B item 6}
  \item{Ab}{Factor A item 1}
  \item{Al}{Factor A item 2}
  \item{Af}{Factor A item 3}
  \item{An}{Factor A item 4}
  \item{Ar}{Factor A item 5}
  \item{Ac}{Factor A item 6}
  \item{sex}{Sex of the child; 1 = boy, 2 = girl}
}
}
\keyword{datasets}
