% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_nl.R
\docType{data}
\name{synthetic_nl}
\alias{synthetic_nl}
\title{Simulated data about morality and politics in The Netherlands}
\format{
A data frame with 518 rows and 33 variables.
}
\usage{
data(synthetic_nl)
}
\description{
This is a simulated counterpart of data presented
by Van Leeuwen and colleagues (2022) concerning
associations between moral dispositions (measured using the Morality As
Cooperation questionnaire, MAC; see Curry et al., 2019) and political
orientation.
}
\references{
Van Leeuwen, F., Van Lissa, C. J., Papakonstantinou, T.,
Petersen, M., & Curry, O. S. (2022, May 25). Morality as Cooperation,
Politics as Conflict. <doi:10.31234/osf.io/wm6rk>

Curry, O. S., Chesters, M. J., & Van Lissa, C. J. (2019). Mapping morality
with a compass: Testing the theory of ‘morality-as-cooperation’with a new
questionnaire. Journal of Research in Personality, 78, 106-124.
}
\keyword{datasets}
