% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_wrapper.R
\name{t_test}
\alias{t_test}
\title{Student's t-test}
\usage{
t_test(x, ...)
}
\arguments{
\item{x}{An object for which an S3 method of t.test exists (vector or
formula).}

\item{\dots}{arguments passed to \code{\link[stats]{t.test}}.}
}
\value{
A list with class \code{"bain_htest"}.
}
\description{
This function is a wrapper for the function \code{\link[stats]{t.test}},
which group-specific sample sizes and variances, in addition to the usual
output of \code{\link[stats]{t.test}}.
}
\details{
This wrapper allows users to enjoy the functionality of bain with the familiar
interface of the stats-function t.test.

For more documentation, see \code{\link[stats]{t.test}}.
}
\examples{
t_test(extra ~ group, data = sleep)
t_test(extra ~ group, data = sleep, paired = TRUE)
t_test(sesamesim$postnumb)
t_test(sesamesim$prenumb, sesamesim$postnumb)
t_test(sesamesim$prenumb, sesamesim$postnumb, paired = TRUE)
}
\seealso{
\code{\link[stats]{t.test}}
}
\keyword{htest}
