% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bain_methods.R
\name{bain}
\alias{bain}
\title{Bayes factors for informative hypotheses}
\usage{
bain(x, hypothesis, ...)
}
\arguments{
\item{x}{An R object containing the outcome of a statistical analysis. Currently, the following
objects can be processed:
\itemize{
\item \code{lm()} objects (anova, ancova, multiple regression). In the Example section it is
elaborated which calls to \code{lm} can be processed by \code{bain} (other calls cannot be processed).
\item \code{t.test()} objects (Student's t-test, Welch's t-test, paired samples t-test, one-group t-test, equivalence test).
\item A named vector containing the estimates resulting from a statistical analysis.
Note that, named means that each estimate has to be labelled such that it can be referred to
in \code{hypotheses}.
}}

\item{hypothesis}{A character string containing the informative hypotheses to evaluate (see Details).}

\item{...}{Additional arguments (see Details).}
}
\value{
The commands \code{bain()} or \code{results<-bain()} followed by \code{results} or \code{print(results)} will
render the default (most important) output from \code{bain}. These concern for each hypothesis specified in \code{hypothesis}
the fit, complexity, Bayes factor versus its complement, posterior model probability (based on equal prior model probabilities)
excluding the unconstrained hypothesis, and posterior model probability including the unconstrained hypothesis. In Hoijtink, Mulder,
van Lissa, and Gu (2018) it is elaborated how these quantities (and the outer output presented below) should be interpreted.
Additionally a descriptives matrix is provided
in which for each estimate, the label, the value, and a 95\% central credibility interval is presented.
The following commands can be used to retrieve the default and additional information from the \code{bain} output object:
\enumerate{
\item \code{results$fit} renders the default output, \code{results$fit$Fit} contains only the column containing the fit of each
hypothesis. In the last command \code{Fit} can be replaced by \code{Com}, \code{BF}, \code{PMPa}, \code{PMPb} to obtain the information in the
corresponding columns of the
default output.
\item \code{results$BFmatrix} contains the matrix containing the mutual Bayes factors of the hypotheses specified in \code{hypotheses}.
\item \code{results$b} contains for each of the groups in the analysis the fraction of information of the data in the group
at hand used to specify the covariance matrix of the prior distribution.
\item \code{results$prior} contains the covariance matrix of the prior distribution.
\item \code{results$posterior} contains the covariance matrix of the posterior distribution.
\item \code{results$call} displays the call to \code{bain}.
\item \code{results$model} displays the named vector or the call to the R function rendering the R object that is input to \code{bain}.
\item \code{results$n} displays the sample sizes per group.
\item \code{results$independent_restrictions} displays the number of independent constraints in the set of hypotheses under
consideration.
\item \code{results$fit$Fit_eq} displays the fit of the equality constrained part of each hypothesis. Replacing \code{Fit_eq} by
\code{Fit_in}, renders the fit of the inequality constrained part of an hypothesis conditional on the fit of the
equality constrained part. \code{Com_eq}, and \code{Com_in}, repectively, are the complexity counterparts of \code{Fit_eq}, and
\code{Fit_in}.
}
}
\description{
\code{bain} is an acronym for "Bayesian informative hypotheses evaluation". It uses the Bayes factor to
evaluate hypotheses specified using equality and inequality constraints among (linear combinations of)
parameters in a wide range of statistical models. A tutorial is provided by Hoijtink, Mulder, van Lissa,
and Gu (2018) retrievable from the Psychological Methods website at \url{https://www.apa.org/pubs/journals/met/}
or the bain website at \url{https://informative-hypotheses.sites.uu.nl/software/bain/} Users are advised to read
this tutorial before using \code{bain}.
}
\details{
========== Using \code{bain} with an \link{lm} or \link{t.test} object ==========

The following steps need to be executed:
\enumerate{
\item \code{x <- lm()} or \code{x <- t.test()}. Execute an analysis with \link{lm} or \link{t.test}.
See Examples for a complete elaboration of the calls to \code{lm} and \code{t.test}
that can be processed by \code{bain}.
\item \code{get_estimates(x)}.	Displays the estimates. In the next step new names are attached to the
estimates. These names are used to specify \code{hypotheses}.
\item \code{label_estimates(x,labels)}. Note that, \code{labels} is a character vector containing new labels
for the estimates in \code{x}. Each label has
to start with a letter, and may consist of "letters", "numbers", ".", and "_".
\item \link{set.seed}\code{(seed)}. Set \code{seed} equal to an integer number to create a repeatable
random number sequence.
\item \code{results <- bain(x,hypotheses)} or \code{bain(x,hypotheses,standardize = TRUE)}. The first call to \code{bain}
is used in case of \code{lm} implementations of anova, ancova, and \code{t.test}. The second call to \code{bain} is used in
case of \code{lm} implementations of multiple regression. With \code{standardize = TRUE} hypotheses with respect to
standardized regression coefficients are evaluated. With \code{standarize = FALSE} hypotheses with respect to
unstandardized regression coefficients are evaluated.
\item \code{print(results)} Print the results of an analysis wity \code{bain}.
\item \code{descriptives(results)} Present descriptives for the parameters used to
specify \code{hypotheses}.
}

========== Using \code{bain} with a named vector ==========

The following steps need to be executed:
\enumerate{
\item Execute a statistical analysis. In case of a single group analysis, the following information has to be
extracted from the statistical analysis and supplied to \code{bain}: 1) a vector containing
estimates of the parameters used to specify \code{hypotheses}; 2) the covariance matrix of these parameters; and, 3)
the sample size. In case of a multiple group analysis, the following information has to be extracted from the
statistical analysis and supplied to \code{bain}: 1) a vector containing estimates of the parameters used to
specify \code{hypotheses} possibly augmented with the estimates of other parameters; per group the covariance
matrix of the parameters corresponding to the group at hand and, possibly, the augmented parameters; and, per group the sample size.
\item Assign names to the estimates
using \code{names(estimates)<-labels}. Note that, \code{labels} is a character vector containing new labels
for the estimates in \code{estimates}. Each label has to start with a letter, and may consist of "letters",
"numbers", ".", and "_". These labels are used to specify \code{hypotheses} (see below).
\item \link{set.seed}\code{(seed)}. Set \code{seed} equal to an integer number to create a repeatable
random number sequence.
\item \code{results <- bain(estimates,hypotheses,n=.,Sigma=.,group_parameters=0,joint_parameters=2)} executes \code{bain} with
the followiong arguments:
\itemize{
\item \code{estimates} A named vector with parameter estimates.
\item \code{hypotheses} A character string containing the informative hypotheses to evaluate (the specification
is elaborated below).
\item \code{n} A vector containing the sample size of each group in the analysis. See, Hoijtink, Gu, and
Mulder (2018), for an elaboration of the difference between one and multiple group analyses. A multiple group
analysis is required when group specific parameters are used to formulate \code{hypotheses}. Examples are the
Student's and Welch's t-test, ANOVA, and ANCOVA. See the Examples section for elaborations of the specification
of multiple group analyses when a named vector is input for \code{bain}.
\item \code{Sigma} In case of a single group analysis (specified using \code{group_parameters = 0}) the
covariance matrix of the parameters. If for a single group analysis \code{joint_parameters = 3} this is a 3 x 3 matrix. In case
of a multiple group analysis a list containing for each group the covariance matrix of the parameters.
If, for example, \code{group_parameters = 1}), there are 3 groups, and \code{joint_parameters = 4}, this is,
a list containing 3 matrices with dimensions 5 x 5.
\item \code{group_parameters} In case of one group \code{group_parameters = 0}. In case of two or more groups,
the number of group specific parameters. In, for example, an ANOVA with three groups, and
\code{joint_parameters = 0},  \code{estimates} will contain three sample means and \code{group_parameters = 1}
because each group is characterized by one mean. In, for example, an ANCOVA with  three groups and
two covariates, \code{estimates} will contain five parameters (three adjusted means and the regression
coefficients of two covariates), \code{group_parameters = 1} because each group is
characterized by one adjusted mean, and \code{joint_parameters = 2} because there are two regression
coefficients that apply to each group. In, for example, a repeated measures design with four repeated measures
and two groups (a between facor with two levels and a  within factor with four levels) \code{estimates} will contain
eight means (four for each group), \code{group_parameters = 4} because
each group is characterized by four means and \code{joint_parameters = 0} because there are no parameters
that apply to each of the groups.
\item \code{joint_parameters} In case of one group the number of parameters in \code{estimates}.
In case of two or more groups, the number of parameters in \code{estimates} shared by the groups. In, for example,
an ANCOVA, the number of \code{joint_parameters} equals the number of covariates.
}
\item \code{print(results)} Print the results of an analysis wity \code{bain}.
\item \code{descriptives(results)} Present descriptives for the parameters used to
specify \code{hypotheses}.
}

========== The specification of \code{hypotheses} ==========

\code{hypotheses} is a character string that specifies which informative hypotheses
have to be evaluated. A simple example is \code{hypotheses <- "a > b > c; a = b = c;"} which
specifies two hypotheses corresponding
to a named vector containing three estimates with names "a", "b", and "c", respectively.

The hypotheses specified have to adhere to the following rules:
\enumerate{
\item Parameters are referred to using the names specified in \code{label_estimates()} or \code{names()}.
\item Linear combinations of parameters can be specified using numbers, +, -, and *. For example, \code{3 * a + 5} or
\code{a + 2 * b + 3 * c}.
\item (Linear combinations of) parameters can be constrained using <, >, and =. For example, \code{a > 0} or
\code{a > b = 0} or \code{2 * a < b + c > 5}.
\item The ampersand & can be used to combine different parts of a hypothesis. For example, \code{a > b & b > c} which
is equivalent to \code{a > b > c} or \code{a > 0 & b > 0 & c > 0}.
\item Sets of (linear combinations of) parameters subjected to the same constraints can be specified using (). For
example, \code{a > (b,c)} which is equivalent to \code{a > b & a > c}.
\item The specification of a hypothesis is completed by typing ; For example, \code{hypotheses <- "a > b > c; a = b = c;"},
specifies two hypotheses.
\item Hypotheses have to be compatible, non-redundant and possible. What these terms mean will be elaborated below.
}

\emph{The set of hypotheses has to be compatible}. For the statistical background of this requirement see Gu, Mulder,
Hoijtink (2018). Usually the sets of hypotheses specified by researchers are compatible, and if not, \code{bain} will return an
error message. The following steps can be used to determine if a set of hypotheses is compatible:
\enumerate{
\item	Replace a range constraint, e.g., \code{1 < a1 < 3}, by an equality constraint in which the parameter involved is equated
to the midpoint of the range, that is, \code{a1 = 2}.
\item Replace in each hypothesis the < and > by =. For example, \code{a1 = a2 > a3 > a4} becomes \code{a1 = a2 = a3 = a4}.
\item The hypotheses are compatible if there is at least one solution to the resulting set of equations. For the two
hypotheses considered under 1. and 2., the solution is a1 = a2 = a3 = a4 = 2. An example of two non-compatible hypotheses
is \code{hypotheses <- "a = 0; a > 2;"} because there is no solution to the equations \code{a=0} and \code{a=2}.
}

\emph{Each hypothesis in a set of hypotheses has to be non-redundant.} A hypothesis is redundant if it can also be specified with
fewer constraints. For example, \code{a = b & a > 0 & b > 0} is redundant because it can also be specified as \code{a = b & a > 0}.
\code{bain} will work correctly if hypotheses specified using only < and > are redundant. \code{bain} will return an error message
if hypotheses specified using at least one = are redundant.

\emph{Each hypothesis in a set of hypotheses has to be possible.} An hypothesis is impossible if estimates in agreement with the
hypothesis do not exist. For example: values for \code{a} in agreement with \code{a = 0 & a > 2} do not exist. It is the
responsibility of the user to ensure that the hypotheses specified are possible. If not, \code{bain} will either return an error
message or render an output table containing many \code{NA}'s.
}
\examples{

\dontrun{
# ==============================================================
# Example Data Set Build into bain
# ==============================================================

# Unless indicated otherwise, the examples that follow below use a simulated data set inspired by the Sesame Street data set from:
# Stevens, J. P. (1996). Applied Multivariate Statistics for the Social Sciences. Mahwah NJ: Lawrence Erlbaum. This data set
# is included in the bain package. The variables contained in sesamesim.txt are subsequently:

# + sex (1 = boy, 2 = girl) of the child
# + site (1 = disadvantaged inner city, 2 = advantaged suburban , 3 = advantaged rural,
#   4 = disadvantaged rural, 5 = disadvantaged Spanish speaking) from which the child originates
# + setting (1 = at home, 2 = at school) in which the child watches sesame street
# + age (in months) of the child
# + viewenc (0 = no, 1 = yes), whether or not the child is encouraged to watch Sesame Street
# + peabody (mental age) score of the child (higher score is higher mental age)
# + prenumb (score on a numbers test before watching Sesame Street for a year)
# + postnumb (score on a numbers test after watching Sesame Street for a year)
# + funumb (follow up numbers test score measured one year after postnumb).

# =================================================================
# The examples that follow below are organized in three categories:
# Running bain with a t.test object; running bain with a lm object;
# and, running bain with a named vector. The ANOVA and ANCOVA
# examples are provided using both an lm object and a named vector
# as input for bain. Below you will find the following examples:
# =================================================================
#
# EXAMPLES USING A T.TEST OBJECT
# + BayesianStudent's t-test (equal within group variances)
# + Bayesian Welch's t-test (unequal within group variances
# + Bayesian paired samples t-test
# + Bayesian one group t-test
# + Bayesian Equivalence test
#
# EXAMPLES USING A LM OBJECT
# + Bayesian ANOVA
# + Bayesian ANCOVA
# + Bayesian multiple regression
#
# EXAMPLES USING A NAMED VECTOR
# + Bayesian ANOVA
# + Bayesian ANCOVA
# + Bayesian repeated measures analysis (one within factor)
# + Bayesian repeated measures analysis (within between design)
# + Bayesian one group logistic regression (counterpart of multiple regression)
# + Bayesian multiple group logistic regression (counterpart of ANCOVA)
# + Bayesian robust ANOVA
# + Bayesian multiple regression with missing data
# + Bayesian structural equation modelling

# ADDITIONAL EXAMPLES CONTAINED IN BFTutorial.R DOWNLOADABLE FROM
# the bain website at \\url{https://informative-hypotheses.sites.uu.nl/software/bain/}
# + Executing a sensitivity analysis with bain
# + Executing Bayesian updating with bain

# ===============================================================================
# An example of Student's t-test (equal within group variances)
# ===============================================================================
# load the bain package which includes the simulated sesamesim.txt data set
library(bain)
# read the data file
sesamesim <- read.table("sesamesim.txt", header=TRUE)
# collect the data for the boys in the vector x and for the girs in the vector y
x<-sesamesim$postnumb[which(sesamesim$sex==1)]
y<-sesamesim$postnumb[which(sesamesim$sex==2)]
# execute student's t-test
ttest <- t.test(x,y,paired = FALSE, var.equal = TRUE)
# inspect the estimates of the group means
get_estimates(ttest)
 # assign names to the estimates
ttest <- label_estimates(ttest, c("boy","girl"))
# set a seed value
set.seed(100)
# test hypotheses with bain
results <- bain(ttest, "boy=girl; boy>girl; boy<girl")
# display the results
print(results)
# obtain the descriptives table
descriptives(results, ci = 0.95)
}


}
\references{
All references are retrievable via or from \url{https://informative-hypotheses.sites.uu.nl/software/bain/}

Bosman, M. and Hoijtink, H. (unpublished). Robust Bayes factors for Bayesian Anova: overcoming
overcoming adverse effect of non-normality and outliers.

Gu, X., Mulder, J., and Hoijtink, H. (2018). Approximate adjusted fractional Bayes factors:
A general method for testing informative hypotheses. \emph{British Journal of Mathematical
and Statistical Psychology, 71,} 229-261. DOI: 10.1111/bmsp.12110

Gu, X., Hoijtink, H., Mulder, J., and Rosseel, Y. (unpublished). Bain: A program for Bayesian
testing of order constrained hypotheses in structural equation models.

Hoijtink, H., Mulder, J., van Lissa, C., and Gu, X. (2018). A tutorial on testing hypotheses
using the Bayes factor. \emph{Psychological Methods.} DOI: 10.1037/met0000201

Hoijtink, H., Gu, X., and Mulder, J. (2018). Bayesian evaluation of informative hypotheses for
multiple populations. \emph{British Journal of Mathematical and Statistical Psychology.}
DOI: 10.1111/bmsp.12145

Hoijtink, H., Gu, X., Mulder, J., and Rosseel, Y. (in press). Computing Bayes Factors
from Data with Missing Values. \emph{Psychological Methods.} DOI: 10.1037/met0000187
}
\seealso{
The bain website at \url{https://informative-hypotheses.sites.uu.nl/software/bain/} contains downloadable legacy
versions of bain (Bain-0.0.1 until Bain-0.1.2). Furthermore, it contains contact information, messages, and announcements
of upcoming workshops. A part of bain (Bayesian t-tests, ANOVA, ANCOVA, and multiple regression) is implemented in
JASP \url{http://jasp-stats.org/}. JASP provides a user-friendly interface and is suited for students and researchers
that are not familiar with the R package.
}
\author{
The main authors of the bain package are Xin Gu (computational core of the code, statistical underpinings),
Caspar van Lissa (user interface and maintainer), Herbert Hoijtink (statistical underpinnings) and Joris Mulder
(statistical underpinnings). Contributions were made by Marlyne Bosman (robust Bayes factors) and Camiel van Zundert
(\code{bain} for structural equation models). Contact information can be found on the bain website at
\url{https://informative-hypotheses.sites.uu.nl/software/bain/}
}
