% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bain_methods.R
\name{bain}
\alias{bain}
\title{Bayes factors for informative hypotheses}
\usage{
bain(x, hypothesis, ...)
}
\arguments{
\item{x}{An R object containing the outcome of a statistical analysis. Currently, the following
objects can be processed:
\itemize{
\item \code{lm()} objects (anova, ancova, multiple regression)
\item \code{t.test()} objects (Student's t-test, Welch's t-test, paired samples t-test, one-group t-test)
\item A named vector containing the estimates resulting from a statistical analysis.
Note that, named means that each estimate has to be labelled such that it can be referred to
in \code{hypotheses}.
}}

\item{...}{Additional arguments (see Details).}

\item{hypotheses}{A character string containing the informative hypotheses to evaluate (see Details).}
}
\value{
Returns a list that contains hypothesis testing results (i.e., Bayes
factors (BFs), relative fit (f), relative complexity (c), posterior model
probabilities (PMPs)), the approximated posterior and prior covariance
matrices, the approximated posteriors means, and the fraction (b).
}
\description{
\code{bain} is an acronym for "Bayesian informative hypothesis evaluation". It uses the Bayes factor to
evaluate hypotheses specified using equality and inequality constraints among (linear combinations of)
parameters in a wide range of statistical models. An introduction is given in Hoijtink, Mulder, van Lissa,
and Gu (2018) retrievable from the Psychological Methods website \url{https://www.apa.org/pubs/journals/met/}
or the bain website \url{https://informative-hypotheses.sites.uu.nl/software/bain/}
}
\details{
========== Using \code{bain} with an \link{lm} or \link{t.test} object ==========

The following steps need to be executed:
\enumerate{
\item \code{x <- lm()} or \code{x <- t.test()}. Execute an analysis with \link{lm} or \link{t.test}.
See Examples for a complete elaboration of analyses that can and cannot be processed by \code{bain}.
\item \code{get_estimates(x)}.	Displays the estimates and the name attached to each estimate. These names are
used to specify \code{hypotheses}.
\item \code{label_estimates(x,labels)}. Note that, \code{labels} is a character vector containing new labels
for the estimates in \code{x}, in order to obtain meaningful and easy to use names. Note that, each name has
to start with a letter, and may consist of "letters", "numbers", ".", and "_".
\item \link{set.seed}\code{(seed)}. Set \code{seed} equal to an integer number to create a repeatable
random number sequence.
\item \code{bain(x,hypotheses)} or \code{bain(x,hypotheses,standardize = TRUE)}. The first call to \code{bain}
is used in case of \code{lm} implementations of anova, ancova, and t.test. The second call to \code{bain} is used in
case of \code{lm} implementations of multiple regression. With \code{standardize = TRUE} hypotheses with respect to
standardized regression coefficients are evaluated. With \code{standarize = FALSE} hypotheses with respect to
unstandardized regression coefficients are evaluated.
}

========== Using \code{bain} with a named vector ==========

The following steps need to be executed:
\enumerate{
\item Execute a statistical analysis. Collect the estimates of interest in a vector. Assign names to the estimates
using \code{names(estimates)<-labels}. Note that, \code{labels} is a character vector containing new labels
for the estimates in \code{estimates}, in order to obtain meaningful and easy to use names. Note that,
each name has to start with a letter, and may consist of "letters", "numbers", ".", and "_".
\item \link{set.seed}\code{(seed)}. Set \code{seed} equal to an integer number to create a repeatable
random number sequence.
\item \code{bain(estimates,hypotheses,n=.,Sigma=.,group_parameters=0,joint_parameters=2)}. Execute \code{bain} with
arguments as elaborated next:
\itemize{
\item \code{estimates} A named vector with parameter estimates.
\item \code{hypotheses} A character string containing the informative hypotheses to evaluate (the specification
is elaborated in the next subsection).
\item \code{n} A vector containing the sample size of each group in the analysis. SOMS SINGLE NUMBER SOMS NIET
\item{Sigma} A list containing per group, the covariance matrix of the parameters (the size of this matrix is
groups+joint_parameters x groups+joint_parameters) SOMS LIST SOMS NIET
\item \code{group_parameters} In case of one group group_parameters = 0. In case of two or more groups,
the number of group specific parameters.
In, for example, an ANOVA with three group, and joint_parameters = 0,  est will contain three
parameters and group_parameters = 1
because each group is characterized by one mean. In, for example, an ANCOVA with  three groups and
two covariates, est will contain
five parameters (three adjusted means and the regression coefficients of two covariates),
group_parameters = 1 because each group is
characterized by one adjusted mean, and joint_parameters = 2 because there are two regression
coefficients that apply to each group.
\item \code{joint_parameters} In case of one group the number of parameters in est. In case of two or more groups, the number of parameters in est
shared by the groups.
}
}
}
\examples{
\dontrun{
#One group:
#Example 1:
#Hypothesis
#H1: theta1>theta2>theta3   #coefficients in regression model

#Input
estimate<-c(3,2,1)  #estimates of coefficients
Sigma<-matrix(c(3,0,0,0,2,0,0,0,1),3,3,byrow = TRUE) #covariance matrix of coefficients

n<-50 #samples size

#H1
ERr<-NULL
IRr<-matrix(c(1,-1,0,0,0,1,-1,0),nrow=2,ncol=4,byrow = TRUE)

res<-Bain(estimate=estimate,grouppara=0,jointpara=3,Sigma=Sigma,n=n,ERr,IRr) #run
#Results are printed.
#Results for fit, complexity, Bayes factor, and PMPs are saved in "res":

plot(res)
#Results for PMPs are plotted.


#Multiple groups
#Example 2
#t test:
#Hypotheses:
#H1: theta1=theta2   #group means
#H2: theta1>theta2
#H3: theta1<theta2

#Input
estimate<-c(0,0)         #Estimates of group means theta1 and theta2.

cov1<-matrix(c(.5),1,1)
cov2<-matrix(c(.1),1,1)
Sigma<-list(cov1,cov2) #List of variances of group means

n<-c(22,37)           #samplesize

#H1:
ERr1<-matrix(c(1,-1,0),nrow=1,ncol=3,byrow = TRUE)
IRr1<-NULL

#H2
ERr2<-NULL
IRr2<-matrix(c(1,-1,0),nrow=1,ncol=3,byrow = TRUE)

#H3
ERr3<-NULL
IRr3<-matrix(c(-1,1,0),nrow=1,ncol=3,byrow = TRUE)

res<-Bain(estimate,Sigma,grouppara=1,jointpara=0,n=n,ERr1,IRr1,ERr2,IRr2,ERr3,IRr3) #run
#Results are printed.
#Results for fit, complexity, Bayes factor, and PMPs are also saved in "res":

plot(res)
#Results for PMPs are plotted.
}
}
\references{
Gu, X., Mulder, J., and Hoijtink, H. (2017). Approximated
adjusted fractional Bayes factors: A general method for testing informative
hypotheses. British Journal of Mathematical and Statistical Psychology.

Hoijtink, H., Gu, X., and Mulder, J. (unpublished). Multiple group Bayesian
evaluation of informative hypotheses.
}
\author{
Caspar van Lissa, Xin Gu, Herbert Hoijtink, Joris Mulder
}
\keyword{htest}
