% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrices.R
\name{constraint_to_equation}
\alias{constraint_to_equation}
\title{Constraint to equation}
\usage{
constraint_to_equation(hyp)
}
\arguments{
\item{hyp}{Character. An (in)equality constraint}
}
\value{
Character
}
\description{
Formats a constraint as an equation that can be evaluated. Adds scalars
to all parameters in the constraint, and adds a XXXconstant parameter to any
constants. Also adds "*" and "+" operators where necessary for evaluation.
}
\keyword{internal}
