% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrices.R
\name{constraint_to_row}
\alias{constraint_to_row}
\title{Constraint to row}
\usage{
constraint_to_row(varnames, hyp)
}
\arguments{
\item{hyp}{Character. An (in)equality constraint formatted as equation,
with all terms on the left hand side.}
}
\value{
Numeric vector.
}
\description{
Evaluate a constraint, which has been formatted as an equation, with
all terms moved to the left hand side, and return a single row of a
(in)equality constraint matrix.
}
\keyword{internal}
