% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrices.R
\name{order_terms}
\alias{order_terms}
\title{Order terms}
\usage{
order_terms(hyp)
}
\arguments{
\item{hyp}{Character. An (in)equality constraint formatted as equation.}
}
\value{
Character
}
\description{
Moves all terms on the right hand side of a constraint, which has been
formatted as an equation, to the left hand side.
}
\keyword{internal}
